move stream codec to sub package and reformat
