+++ src/main/java/com/almostreliable/summoningrituals/network/codec/package-info.java
@@ -1,0 +1,4 @@
+@NonnullDefault
+package com.almostreliable.summoningrituals.network.codec;
+
+import org.lwjgl.system.NonnullDefault;
